second

+++ src/pages/Home.js
@@ -1,12 +1,13 @@
 import axios from "axios";
 import React, { useEffect, useState } from "react";
 import { Button } from "react-bootstrap";
-import '../style/home.css'
+import '../style/home.css';
+import Swal from 'sweetalert2';
 
 export default function Home() {
   const [buku, setBuku] = useState([]);
 
-  const getAll = () => {
+  const getAll = () => { //menampilkan semua data
     axios
       .get("http://localhost:8000/daftarBuku")
       .then((res) => {
@@ -17,13 +18,30 @@ export default function Home() {
       });
   };
 
-  useEffect(() => {
+  useEffect(() => { //mengaktifkan method
     getAll();
   }, []);
-
-  const deleteUser = async (id) =>  {
-    axios.delete("http://localhost:8000/daftarBuku/" + id);
-    alert("User berhasil dihapus ges..");
+  
+  const deleteUser = async (id) =>  { //method delete
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        axios.delete("http://localhost:8000/daftarBuku/" + id);
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+          )
+          window.location.reload();
+      }
+    })
     getAll();
   };
 
@@ -37,7 +55,7 @@ export default function Home() {
             <th>Deskripsi</th>
             <th>Pengarang</th>
             <th>Tahun Terbit</th>
-            <th>Aksi</th>
+            {localStorage.getItem("id") !== null ? <th>Aksi</th> : <></>}
           </tr>
         </thead>
         <tbody>
@@ -48,20 +66,22 @@ export default function Home() {
               <td>{book.deskripsi}</td>
               <td>{book.pengarang}</td>
               <td>{book.tahunTerbit}</td>
+              {localStorage.getItem("id") !== null ? (
               <td>
                 <Button
                   variant="danger"
                   className="mx-1"
                   onClick={() => deleteUser(book.id)}
-                >
-                  Hapus
-                </Button>
-                <a href={"/edit/" + book.id}>
-                  <Button variant="warning" className="mx-1">
-                    Ubah
+                  >
+                   Hapus
                   </Button>
-                </a>
-              </td>
+                  <a href={"/edit/" + book.id} > {/* beda page*/}
+                    <Button variant="warning" className="mx-1">
+                      Ubah
+                    </Button>
+                  </a>
+              </td> 
+              ) : (<></>)}
             </tr>
           ))}
         </tbody>
